requirements.txt
numpy
pandas>=1.3.0  # Ensures a version of Pandas 1.3.0 or newer
torch>=1.7,<=2.0  # Allows for flexibility in the PyTorch version
torch-geometric==2.3.0
networkx==3.1
scikit-learn
matplotlib==3.7.1
seaborn==0.12.1
hydra-core==1.3.2
pyyaml==6.0
nltk
PyYAML